backend/requirements.txt
# FastAPI and Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# HTTP Client
httpx>=0.27.0

# Security
cryptography>=42.0.0
pyjwt>=2.8.0

# Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Affilync Shared Library
affilync-integrations-common @ file:///home/ninja/codex/affilync-integrations-common

# Development/Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
